refactor(confirmations): move universal transaction rows
