+++ ui/pages/confirmations/components/rows/solana/solana-test-count-row.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { ConfirmInfoRow } from '../../../../../../../components/app/confirm/info/row/row';
-import { useSolanaTransactionData } from '../../../../../hooks/transactions/useSolanaTransactionData';
+import { ConfirmInfoRow } from '../../../../../components/app/confirm/info/row/row';
+import { useSolanaTransactionData } from '../../../hooks/transactions/useSolanaTransactionData';
 
 export function SolanaTestCountRow() {
   const data = useSolanaTransactionData();
