+++ ui/pages/confirmations/components/rows/universal/truncate-universal-address.ts
